possibly fix spurious error from cust_bill.pm related to barcode files when barcodes are turned off

+++ FS/FS/cust_bill.pm
@@ -3344,7 +3344,7 @@ sub print_ps {
   my ($file, $logofile, $barcodefile) = $self->print_latex(@_);
   my $ps = generate_ps($file);
   unlink($logofile);
-  unlink($barcodefile);
+  unlink($barcodefile) if $barcodefile;
 
   $ps;
 }
@@ -3373,7 +3373,7 @@ sub print_pdf {
   my ($file, $logofile, $barcodefile) = $self->print_latex(@_);
   my $pdf = generate_pdf($file);
   unlink($logofile);
-  unlink($barcodefile);
+  unlink($barcodefile) if $barcodefile;
 
   $pdf;
 }
